move session creation
